NEW: Add MutableSource and MutableString.
These are 2 more support classes that make it easier to build upgraders
that make minimal changes to the code being modified.

MutableString is a support class that will collect a number of string
replacements made with consistnet reference to the character positions
of the original string.

MutableSource builds on this to provide support methods that are aware
of the PHP-Parser’s AST.