FIX: A couple more edge-cases in class renaming.
Static consts & vars weren’t getting recognised, this fixes that. A
scan of the available AST nodes in php-parser suggests that this is the
last of them.

Trait uses were also not getting renamed.

The actual implementations of classes, interfaces, and traits are *not*
affected by these, but that seems beyond the scope of a tool for
updating project code in response to changes made within libraries by
developers.