NEW: Add namespace correction to RenameClasses
RenameClasses now has a feature designed to help with developers who
have added a namespace to their class but haven’t corrected the class
references within it.

It’s probably best refactored into a separate subclass of RenameClasses.
Later, this kind of functionality may be better accessed with a separate
refactoring command rather than being chucked into the upgrade tool.

For now, it’s useful for helping to test.