Inspection capabilities - warnings instead of code rewrites
Given the sheer number of small API changes, it'll be nearly impossible
for developers to proactively upgrade their codebase
since it would mean keeping 100+ API changes in their heads
while looking through each of their project files.

We can't fully automate this process,
since PHP parsers don't know enough about the relationships.
But we can get close: If you're invoking a manyManyComponent() method
somewhere, there's a strong likelyhood that you're doing it on a DataObject
or subclass, due to the unique method naming. We don't need to get the
exact class to give you a recommendation.